extract commandNameAndArgs into independent function
